requirements.txt
fastapi
uvicorn
tensorflow==2.19.0
keras==3.10.0
numpy
pillow
python-multipart